test: name test classes for behaviour and drop low-value tests
Rename test classes after the behaviour they verify rather than the command or
class they were carved out of, split classes that covered unrelated behaviours,
and remove tests that barely exercised anything.

- Renamed 10 single-purpose classes (e.g. AdminCommandsTest -> BlockedConfigKeyTest,
  NominationServiceTest -> PromotionNominationDecisionTest, PunishmentServiceTest ->
  PunishmentSearchAndStatsTest).
- Split ProfileCommandsTest -> SocialMediaVerificationTest + BadgeDisplayOrderTest;
  ProfileUpdateFeatureTest -> StaleProfileSelectionTest + NicknameChangeDecisionTest;
  GeneratorCommandsTest -> GeneratedAttachmentFormatTest, PowerStoneCommandStringTest,
  DialogueBuildingTest, PowerStoneStatParsingTest, SearchResultRenderingTest.
- Cut low-value tests: the empty-store purge case, the four null-safe getCommandHistory
  getters, and the redundant attachment base-name test.

ResourcePackServiceTest (heavy shared fixtures) and GlyphListingTest (one focused
class) are left intact - each is a cohesive single-subject suite where splitting
would duplicate fixtures rather than clarify.
